fix net_localgroup tests
